requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-multipart==0.0.5
torch==2.2.0
torchvision==0.17.0
numpy>=1.24.0
Pillow>=9.0.0
python-dotenv==0.19.0 